Reorganized roles/capabilities, added NON_EMPTY_ARRAY
