+++ layers/Schema/packages/user-roles/src/FieldResolvers/RolesFieldResolverTrait.php
@@ -5,10 +5,8 @@ declare(strict_types=1);
 namespace PoPSchema\UserRoles\FieldResolvers;
 
 use PoP\ComponentModel\Schema\SchemaDefinition;
-use PoP\ComponentModel\Schema\SchemaTypeModifiers;
 use PoP\ComponentModel\TypeResolvers\TypeResolverInterface;
 use PoP\Translation\Facades\TranslationAPIFacade;
-use PoPSchema\UserRoles\Facades\UserRoleTypeDataResolverFacade;
 
 trait RolesFieldResolverTrait
 {
@@ -37,17 +35,6 @@ trait RolesFieldResolverTrait
         return $types[$fieldName] ?? parent::getSchemaFieldType($typeResolver, $fieldName);
     }
 
-    public function getSchemaFieldTypeModifiers(TypeResolverInterface $typeResolver, string $fieldName): ?int
-    {
-        return match($fieldName) {
-            'roles',
-            'capabilities'
-                => SchemaTypeModifiers::NON_NULLABLE | SchemaTypeModifiers::IS_ARRAY,
-            default
-                => parent::getSchemaFieldTypeModifiers($typeResolver, $fieldName),
-        };
-    }
-
     public function getSchemaFieldDescription(TypeResolverInterface $typeResolver, string $fieldName): ?string
     {
         $translationAPI = TranslationAPIFacade::getInstance();
@@ -56,31 +43,5 @@ trait RolesFieldResolverTrait
             'capabilities' => $translationAPI->__('All user capabilities', 'user-roles'),
         ];
         return $descriptions[$fieldName] ?? parent::getSchemaFieldDescription($typeResolver, $fieldName);
-    }
-
-    /**
-     * @param array<string, mixed> $fieldArgs
-     * @param array<string, mixed>|null $variables
-     * @param array<string, mixed>|null $expressions
-     * @param array<string, mixed> $options
-     */
-    public function resolveValue(
-        TypeResolverInterface $typeResolver,
-        object $resultItem,
-        string $fieldName,
-        array $fieldArgs = [],
-        ?array $variables = null,
-        ?array $expressions = null,
-        array $options = []
-    ): mixed {
-        $userRoleTypeDataResolver = UserRoleTypeDataResolverFacade::getInstance();
-        switch ($fieldName) {
-            case 'roles':
-                return $userRoleTypeDataResolver->getRoleNames();
-            case 'capabilities':
-                return $userRoleTypeDataResolver->getCapabilities();
-        }
-
-        return parent::resolveValue($typeResolver, $resultItem, $fieldName, $fieldArgs, $variables, $expressions, $options);
     }
 }
