+++ layers/Schema/packages/user-roles/src/FieldResolvers/RootRolesFieldResolver.php
@@ -4,9 +4,12 @@ declare(strict_types=1);
 
 namespace PoPSchema\UserRoles\FieldResolvers;
 
-use PoP\Engine\TypeResolvers\RootTypeResolver;
-use PoPSchema\UserRoles\FieldResolvers\RolesFieldResolverTrait;
 use PoP\ComponentModel\FieldResolvers\AbstractDBDataFieldResolver;
+use PoP\ComponentModel\Schema\SchemaTypeModifiers;
+use PoP\ComponentModel\TypeResolvers\TypeResolverInterface;
+use PoP\Engine\TypeResolvers\RootTypeResolver;
+use PoPSchema\UserRoles\Facades\UserRoleTypeDataResolverFacade;
+use PoPSchema\UserRoles\FieldResolvers\RolesFieldResolverTrait;
 
 class RootRolesFieldResolver extends AbstractDBDataFieldResolver
 {
@@ -17,5 +20,47 @@ class RootRolesFieldResolver extends AbstractDBDataFieldResolver
         return [
             RootTypeResolver::class,
         ];
+    }
+
+    public function getSchemaFieldTypeModifiers(TypeResolverInterface $typeResolver, string $fieldName): ?int
+    {
+        return match($fieldName) {
+            'roles'
+                => SchemaTypeModifiers::NON_NULLABLE
+                | SchemaTypeModifiers::IS_ARRAY
+                | SchemaTypeModifiers::NON_EMPTY_ARRAY,
+            'capabilities'
+                => SchemaTypeModifiers::NON_NULLABLE
+                | SchemaTypeModifiers::IS_ARRAY
+                | SchemaTypeModifiers::NON_EMPTY_ARRAY,
+            default
+                => parent::getSchemaFieldTypeModifiers($typeResolver, $fieldName),
+        };
+    }
+
+    /**
+     * @param array<string, mixed> $fieldArgs
+     * @param array<string, mixed>|null $variables
+     * @param array<string, mixed>|null $expressions
+     * @param array<string, mixed> $options
+     */
+    public function resolveValue(
+        TypeResolverInterface $typeResolver,
+        object $resultItem,
+        string $fieldName,
+        array $fieldArgs = [],
+        ?array $variables = null,
+        ?array $expressions = null,
+        array $options = []
+    ): mixed {
+        $userRoleTypeDataResolver = UserRoleTypeDataResolverFacade::getInstance();
+        switch ($fieldName) {
+            case 'roles':
+                return $userRoleTypeDataResolver->getRoleNames();
+            case 'capabilities':
+                return $userRoleTypeDataResolver->getCapabilities();
+        }
+
+        return parent::resolveValue($typeResolver, $resultItem, $fieldName, $fieldArgs, $variables, $expressions, $options);
     }
 }
